+++ src/main/scala/api/sampling/evaluators/ModelPriorEvaluator.scala
@@ -22,7 +22,7 @@ import scalismo.sampling.DistributionEvaluator
 import scalismo.statisticalmodel.{MultivariateNormalDistribution, StatisticalMeshModel}
 
 case class ModelPriorEvaluator(model: StatisticalMeshModel) extends DistributionEvaluator[ModelFittingParameters] {
-  val mvnormal = MultivariateNormalDistribution(DenseVector.zeros[Double](model.rank), diag(DenseVector.ones[Double](model.rank)))
+  val mvnormal: MultivariateNormalDistribution = MultivariateNormalDistribution(DenseVector.zeros[Double](model.rank), diag(DenseVector.ones[Double](model.rank)))
 
 
   override def logValue(theta: ModelFittingParameters): Double = {
